bloodhound: Add base Bottlerocket checker binary
This adds a multicall binary to be used as the common entry point for
all Bottlerocket CIS benchmark checks.

Bloodhound will discover checkers to run by looking in the designated
directory for any binaries that implement the checker interface. This
makes it easy to add or test additional checks, but there is a slight
overhead to each binary. To balance this, most (or all) Bottlerocket
checks will be routed as symlinks through a common binary. The
bottlerocket-checks binary looks at arg[0] to determine which checker to
run.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>